refactor(frontend): add generic iframe-based ToolPreview
